Start working on descriptor set abstraction
The idea is to only describe the descriptor set once, in a declarative
way: the Gpu is then going to allocate the descriptor set using an
implementor of the DescriptorSetAllocator trait, setup the descriptor
set state and then return it: the user only has to supply the descriptor
set layout.

The current allocator implementation simply created a descriptor set
with a fixed amount of descriptors: when an allocation fails, a new pool
is created.