Use builder pattern for render pass definition
By using the builder pattern for defining render passes
into the render graph, we are forced to call commit, since
otherwise the render graph cannot be used, as it is borrowed
during the render pass definition.

Using the builder pattern also seemed like the most logical choice,
because all the operations to define a render graph are written in a
single code block.