Revert changes to use shadow cubemaps
It appears that the technique i'm using to render shadows in a single
image is valid, it's called Shadow Atlas and it's a technique used in
multiple games.
Given this findings, i made explicit that i'm using a Shadow Atlas, and
i will focusing my efforts on improving my implementation.