Add support for cvar system and console
The console can take inputs, parse them in some way, and then spit out
some output.
The console is accessible through the F12 button

The CvarManager can be used to define global console variables, which
can then be modified and queried. These variables are often used to
configure the application, and can be stored in a config file, which is
then loaded at runtime.

For now, only a cvar named 'g_test' is available