Expose get_ref and get_ref_mut methods to CvarManager
These methods are useful to edit a value stored inside the cvar manager.

In this case, the fxaa iterations have been exposed as a cvar, and are
now edited from the cvar manager.

These methods are the only ones to panic when the type of the cvar is
different from the type of the reference, as casting is not possible