Make ResourceMap Send + Sync
by moving the owrnership of the reference counter to the map itself, +
adding an update method, the resource map has been made Send + Sync.

Every time a refcount increment/decrement is needed, a boxed operation
is sent that is then executed by the resource map.

Also, the tests had to be updated accordingly.