Start implementing pipeline barriers across queues
This change should allow us to start implementing async compute
dispatching.

When a resource (an image or a buffer) is used by a queue, and it
is needed on another queue, a pipeline barrier across queues
must be inserted.
This is done by remebering which queue owns a certain resource when
dispatching compute/executing draw calls

In the future this change might allow for having exclusive queue
ownership of resources